Temporarily solve the warnings related to longValues() being deprecated and some methods not being marked as override.

+++ src/veins/modules/application/platooning/UnicastProtocol.cc
@@ -45,8 +45,8 @@ void UnicastProtocol::initialize(int stage)
         upperControlOut = findGate("upperControlOut");
 
         // get parameters
-        queueSize = par("queueSize").longValue();
-        maxAttempts = par("maxAttempts").longValue();
+        queueSize = par("queueSize").intValue();
+        maxAttempts = par("maxAttempts").intValue();
         ackTimeout = par("ackTimeout").doubleValue();
 
         // timeout message
