+++ src/veins/modules/application/platooning/traffic/PlatoonsTrafficManager.cc
@@ -26,9 +26,9 @@ void PlatoonsTrafficManager::initialize(int stage)
 
     if (stage == 0) {
 
-        nCars = par("nCars").longValue();
-        platoonSize = par("platoonSize").longValue();
-        nLanes = par("nLanes").longValue();
+        nCars = par("nCars").intValue();
+        platoonSize = par("platoonSize").intValue();
+        nLanes = par("nLanes").intValue();
         platoonInsertTime = SimTime(par("platoonInsertTime").doubleValue());
         platoonInsertSpeed = par("platoonInsertSpeed").doubleValue();
         platoonInsertDistance = par("platoonInsertDistance").doubleValue();
